# Minor cleanups for legacy beatmap decoders