@peppy 2020-02-08
When making alloc changes, provding BenchmarkDotNet memory analysis (`[MemoryDiagnoser]`) results is very appreciated (so we don't have to double-check ourselves).