@bdach 2020-02-08
Could include the benchmark in the PR as part of the recently-added [benchmark project](https://github.com/ppy/osu/tree/master/osu.Game.Benchmarks) too, so that anyone can just run it at their convenience. There's already a beatmap parsing benchmark in there, but it seems to be more CPU-oriented.