@turbedi 2020-02-08
Tested with the storyboard of [this beatmap](https://osu.ppy.sh/beatmapsets/470977).

master:
|                     Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|--------------------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| BenchmarkStoryboard | 1.257 s | 0.0064 s | 0.0059 s | 72000.0000 | 20000.0000 | 1000.0000 | 427.92 MB |

This PR:
|                     Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|--------------------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| BenchmarkStoryboard | 1.227 s | 0.0076 s | 0.0067 s | 71000.0000 | 20000.0000 | 1000.0000 | 418.15 MB |

I didn't include the benchmark in this PR, because there aren't any storyboard files under osu.Game.Tests/Resources, [refering to this comment from the benchmark project PR](https://github.com/ppy/osu/pull/7474#discussion_r364533606).