@bdach 2020-02-08
Mind linking the benchmark source as a gist or something at least, then?