@turbedi 2020-02-08
> 
> 
> Mind linking the benchmark source as a gist or something at least, then?

https://gist.github.com/turbedi/eaf4a6e99d0fa6f49acb1f2c2f917b41